Upgrade UIforETW.sln to VS 2015
This upgrades UIforETW.sln to build with VS 2015. It includes fixes for
two warnings. The warnings didn't represent real bugs, but the fix is
easy enough with VC++ 2015's support for %z.

However!

With VC++ 2015 RTM the %zu code triggers /analyze warnings, because
apparently /analyze wasn't updated to know about %zu.

And, due to annotation errors in swprintf_s, most uses of this function
trigger bogus /analyze warnings. All these false-positives are shown
here:
https://github.com/google/UIforETW/issues/58#issuecomment-152851953

Update 1 RC seems to have even worse problems with /analyze. But, the
main compiles work so I am provisionally upgrading in order to simplify
reporting these bugs.

+++ UIforETW/UIforETWDlg.cpp
@@ -328,7 +328,7 @@ BOOL CUIforETWDlg::OnInitDialog()
 					}
 					else
 					{
-						outputPrintf(L"Failure code %d while installing WPT 8.1.\n", reinterpret_cast<int>(installResult81));
+						outputPrintf(L"Failure code %zu while installing WPT 8.1.\n", reinterpret_cast<size_t>(installResult81));
 					}
 				}
 			}
@@ -346,7 +346,7 @@ BOOL CUIforETWDlg::OnInitDialog()
 				}
 				else
 				{
-					outputPrintf(L"Failure code %d while installing WPT 10.\n", reinterpret_cast<int>(installResult10));
+					outputPrintf(L"Failure code %zu while installing WPT 10.\n", reinterpret_cast<size_t>(installResult10));
 				}
 			}
 		}
